+++ Classes/TBODeveloperOverlayLogger/TBODeveloperOverlayLoggerCocoalumberjackDatasource.m
@@ -15,6 +15,8 @@
 #import "NSString+RegexMatches.h"
 #import "NSMutableAttributedString+RegexMatches.h"
 
+#import "TBODeveloperOverlayFilterableLogFormatterProtocol.h"
+
 @interface TBODeveloperOverlayLoggerCocoaLumberjackDatasource ()
 
 @property (strong, nonatomic, readwrite) NSMutableSet<NSRegularExpression *> *logLevelRegexes;
@@ -25,7 +27,7 @@
 @implementation TBODeveloperOverlayLoggerCocoaLumberjackDatasource
 
 - (NSString *)lastLogMessagesLimitedToCharacterCount:(NSUInteger)maxCharacterCount {
-    if (!self.fileLogger){
+    if (!self.fileLogger) {
         return nil;
     }
     NSString *fullLog = [TBODeveloperOverlayLogFileReader lastLogMessagesLimitedToCharacterCount:maxCharacterCount inFilelogger:[self fileLogger]];
@@ -33,13 +35,13 @@
 }
 
 - (NSAttributedString *)attributedLastLogMessagesLimitedToCharacterCount:(NSUInteger)maxCharacterCount {
-    if (!self.fileLogger){
+    if (!self.fileLogger) {
         return nil;
     }
     NSString *filteredLog = [self lastLogMessagesLimitedToCharacterCount:maxCharacterCount];
     NSMutableAttributedString *attributedString = [[NSMutableAttributedString alloc] initWithString:filteredLog];
     if (self.searchString) {
-        NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:[NSString stringWithFormat:@"%@", self.searchString] options:NSRegularExpressionDotMatchesLineSeparators error:nil];
+        NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:[NSString stringWithFormat:@"%@", self.searchString] options:NSRegularExpressionDotMatchesLineSeparators|NSRegularExpressionCaseInsensitive error:nil];
         [attributedString addAttributes:@{NSBackgroundColorAttributeName:[UIColor colorWithRed:0.875 green:1.000 blue:0.886 alpha:1.000]} toMatchesOfRegex:regularExpression];
     }
     return attributedString;
@@ -71,16 +73,35 @@
     }
 }
 
+- (NSRegularExpression *)singleMessageRegularExpression {
+    NSRegularExpression *singleMessageRegularExpression = nil;
+    if ([[[self fileLogger] logFormatter] respondsToSelector:@selector(logMessageRegularExpression)]) {
+        singleMessageRegularExpression = [[[self fileLogger] logFormatter] performSelector:@selector(logMessageRegularExpression)];
+    }
+    if (!singleMessageRegularExpression) {
+        singleMessageRegularExpression = [NSRegularExpression regularExpressionWithPattern:@"^.*$\n" options:NSRegularExpressionAnchorsMatchLines error:nil];
+    }
+    return singleMessageRegularExpression;
+}
+
 - (NSString *)filteredLog:(NSString *)logString {
-    NSString *filteredString = logString;
-    if (self.logLevelRegexes.count > 0) {
-        filteredString = [logString substringThatMatchesAnyRegex:self.logLevelRegexes.allObjects];
-    }
-    if (self.searchString.length > 0) {
-        NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:[NSString stringWithFormat:@".*%@.*", self.searchString] options:kNilOptions error:nil];
-        filteredString = [filteredString substringThatMatchesRegex:regularExpression];
-    }
-    return filteredString;
+    NSMutableArray *filteredMessages = [NSMutableArray new];
+    NSRegularExpression *singleMessageRegularExpression = [self singleMessageRegularExpression];
+    [singleMessageRegularExpression enumerateMatchesInString:[logString stringByAppendingString:@"\n"] options:kNilOptions range:NSMakeRange(0, logString.length) usingBlock:^(NSTextCheckingResult *_Nullable result, NSMatchingFlags flags, BOOL *_Nonnull stop) {
+        NSString *match = [logString substringWithRange:result.range];
+        BOOL matches = true;
+        if (self.logLevelRegexes.count > 0) {
+            matches &= [match matchesEveryRegex:self.logLevelRegexes.allObjects];
+        }
+        if (self.searchString.length > 0) {
+            NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:[NSString stringWithFormat:@".*%@.*", self.searchString] options:NSRegularExpressionCaseInsensitive error:nil];
+            matches &= [match matchesRegex:regularExpression];
+        }
+        if (matches) {
+            [filteredMessages addObject:match];
+        }
+    }];
+    return [filteredMessages componentsJoinedByString:@"\n"];
 }
 
 #pragma mark helper