+++ TBODeveloperOverlayTests/NSString+RegexMatchesTests.m
@@ -2,12 +2,11 @@
 //  NSString+RegexMatchesTests.m
 //  TBODeveloperOverlay
 //
-//  Created by Bernhard Eiling on 14.06.16.
+//  Created by Cornelius Horstmann on 11.11.16.
 //  Copyright © 2016 TBO INTERACTIVE. All rights reserved.
 //
 
 #import <XCTest/XCTest.h>
-
 #import "NSString+RegexMatches.h"
 
 @interface NSString_RegexMatchesTests : XCTestCase
@@ -18,59 +17,19 @@
 
 NSString *testString = @"this is a string\nwhich includes a couple\nof newlines";
 
-- (void)testRegexMatchingSubstring {
-    NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:@"which includes" options:kNilOptions error:nil];
-    NSString *subString = [testString substringThatMatchesRegex:regularExpression];
-    XCTAssertEqualObjects(@"which includes a couple", subString);
+- (void)testMatchesRegex {
+    NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:@"[\\sa-zA-Z]+" options:kNilOptions error:nil];
+    XCTAssertTrue([@"This is a test string" matchesRegex:regularExpression]);
+    XCTAssertTrue([@"923875928374928374 Test for substring matching" matchesRegex:regularExpression]);
+    XCTAssertFalse([@"9879" matchesRegex:regularExpression]);
 }
 
-- (void)testRegexNotMatchingSubstring {
-    NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:@"not matching" options:kNilOptions error:nil];
-    NSString *subString = [testString substringThatMatchesRegex:regularExpression];
-    XCTAssertEqualObjects(@"", subString);
-}
-
-- (void)testAnyRegexMatchingSubstring {
-    NSRegularExpression *regularExpression1 = [NSRegularExpression regularExpressionWithPattern:@"this is" options:kNilOptions error:nil];
-    NSRegularExpression *regularExpression2 = [NSRegularExpression regularExpressionWithPattern:@"which includes" options:kNilOptions error:nil];
-    NSString *subString = [testString substringThatMatchesAnyRegex:@[regularExpression1, regularExpression2]];
-    XCTAssertEqualObjects(@"this is a string\nwhich includes a couple", subString);
-}
-
-- (void)testNotAnyRegexMatchingSubstring {
-    NSRegularExpression *regularExpression1 = [NSRegularExpression regularExpressionWithPattern:@"not matching" options:kNilOptions error:nil];
-    NSRegularExpression *regularExpression2 = [NSRegularExpression regularExpressionWithPattern:@"not matching as well" options:kNilOptions error:nil];
-    NSString *subString = [testString substringThatMatchesAnyRegex:@[regularExpression1, regularExpression2]];
-    XCTAssertEqualObjects(@"", subString);
-}
-
-- (void)testAnyRegexMatchingWithFalseTypes {
-    NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:@"which includes" options:kNilOptions error:nil];
-    NSString *subString = [testString substringThatMatchesAnyRegex:@[regularExpression, @1]];
-    XCTAssertEqualObjects(@"which includes a couple", subString);
-}
-
-- (void)testRegexWithNewline {
-    NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:@"string\nwhich" options:kNilOptions error:nil];
-    NSString *subString = [testString substringThatMatchesRegex:regularExpression];
-    XCTAssertEqualObjects(@"", subString);
-}
-
-- (void)testRegexAboveTwoLines {
-    NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:@"string which" options:kNilOptions error:nil];
-    NSString *subString = [testString substringThatMatchesRegex:regularExpression];
-    XCTAssertEqualObjects(@"", subString);
-}
-
-- (void)testLongRegexDescribingTargetStringTwice {
-    NSRegularExpression *regularExpression = [NSRegularExpression regularExpressionWithPattern:@"this is a string\nwhich includes a couple\nof newlinesthis is a string\nwhich includes a couple\nof newlines" options:kNilOptions error:nil];
-    NSString *subString = [testString substringThatMatchesRegex:regularExpression];
-    XCTAssertEqualObjects(@"", subString);
-}
-
-- (void)testNilRegex {
-    NSString *subString = [testString substringThatMatchesRegex:nil];
-    XCTAssertEqualObjects(@"this is a string\nwhich includes a couple\nof newlines", subString);
+- (void)testMatchesEveryRegex {
+    NSRegularExpression *firstRegularExpression = [NSRegularExpression regularExpressionWithPattern:@"a.*" options:kNilOptions error:nil];
+    NSRegularExpression *secondRegularExpression = [NSRegularExpression regularExpressionWithPattern:@"b.*" options:kNilOptions error:nil];
+    NSArray *expressions = @[firstRegularExpression, secondRegularExpression];
+    XCTAssertTrue([@"abcde" matchesEveryRegex:expressions]);
+    XCTAssertFalse([@"acde" matchesEveryRegex:expressions]);
 }
 
 @end